+++ errpref/errprefdtonanobot.go
@@ -27,6 +27,113 @@ func (ePrefDtoNanobot errPrefixDtoNanobot) ptr() *errPrefixDtoNanobot {
 	}
 }
 
+// setFromIBuilder - Receives a pointer to an ErrPrefixDto instance
+// plus an object implementing the IBuilderErrorPrefix interface.
+//
+// This method then proceeds to copy an transfer all error prefix
+// and error context information from the IBuilderErrorPrefix
+// object to the ErrPrefixDto instance.
+//
+// IMPORTANT
+// All existing error prefix and error context information in this
+// ErrPrefixDto instance will be overwritten and deleted.
+//
+//
+// ----------------------------------------------------------------
+//
+// Input Parameters
+//
+//  errPrefDto          *ErrPrefixDto
+//     - A pointer to an instance of ErrPrefixDto. Data values in
+//       this ErrPrefixDto object will be overwritten an set to new
+//       values extracted from input parameter, 'iEPref'.
+//
+//
+//  iEPref              IBuilderErrorPrefix
+//     - An object which implements the IBasicErrorPrefix
+//       interface.
+//
+//       This interface contains a method named,
+//       'GetEPrefStrings()'. This method returns a two dimensional
+//       array of strings. Each 2-Dimensional array element holds a
+//       separate string for error prefix and error context.
+//
+//       Information extracted from this object will generate error
+//       prefix information used to overwrite and replace the error
+//       prefix and error context information in the input
+//       parameter, 'errPrefDto'.
+//
+//
+//  errorPrefStr        string
+//     - This parameter contains an error prefix string which is
+//       included in all returned error messages. Usually, it
+//       contains the names of the calling method or methods.
+//
+//       If no error prefix information is needed, set this
+//       parameter to an empty string: "".
+//
+//
+// -----------------------------------------------------------------
+//
+// Return Values
+//
+//  error
+//     - If this method completes successfully, the returned error
+//       Type is set equal to 'nil'.
+//
+//       If errors are encountered during processing, the returned
+//       error Type will encapsulate an error message. This
+//       returned error message will incorporate the method chain
+//       and text passed by input parameter, 'errorPrefStr'. The
+//       'errorPrefStr' text will be attached to the beginning of
+//       the error message.
+//
+func (ePrefDtoNanobot *errPrefixDtoNanobot) setFromIBuilder(
+	errPrefDto *ErrPrefixDto,
+	iEPref IBuilderErrorPrefix,
+	errorPrefStr string) error {
+
+	if ePrefDtoNanobot.lock == nil {
+		ePrefDtoNanobot.lock = new(sync.Mutex)
+	}
+
+	ePrefDtoNanobot.lock.Lock()
+
+	defer ePrefDtoNanobot.lock.Unlock()
+
+	methodName := errorPrefStr + "\nerrPrefixDtoNanobot." +
+		"setFromIBuilder()"
+
+	if errPrefDto == nil {
+		return fmt.Errorf("%v\n"+
+			"Error: Input parameter 'errPrefDto' is invalid!\n"+
+			"'errPrefDto' is a 'nil' pointer.\n",
+			methodName)
+	}
+
+	if iEPref == nil {
+		return fmt.Errorf("%v\n"+
+			"Error: Input parameter 'iEPref' is invalid!\n"+
+			"'iEPref' is a 'nil' pointer.\n",
+			methodName)
+	}
+
+	err := errPrefQuark{}.ptr().emptyErrPrefInfoCollection(
+		errPrefDto,
+		methodName)
+
+	if err != nil {
+		return err
+	}
+
+	err = errPrefAtom{}.ptr().addTwoDimensionalStringArray(
+		errPrefDto,
+		iEPref.GetEPrefStrings(),
+		methodName)
+
+	return err
+}
+
 // setFromIBasicErrorPrefix - Receives an ErrPrefixDto object and
 // then proceeds to overwrite and reset the ErrorPrefixInfo
 // collection containing error prefix and error context
@@ -37,10 +144,10 @@ func (ePrefDtoNanobot errPrefixDtoNanobot) ptr() *errPrefixDtoNanobot {
 // The IBasicErrorPrefix interface contains the method:
 //  GetEPrefStrings() [][2]string
 //
-// This method returns a two-dimensional slice of error prefix and
-// error context strings which can be used to create new
-// ErrorPrefixInfo collection. This new ErrorPrefixInfo collection
-// is assigned to the ErrPrefixDto object.
+// IMPORTANT
+// All existing error prefix and error context information in this
+// ErrPrefixDto instance (errPrefDto) will be overwritten and
+// deleted.
 //
 //
 // ----------------------------------------------------------------
@@ -62,9 +169,10 @@ func (ePrefDtoNanobot errPrefixDtoNanobot) ptr() *errPrefixDtoNanobot {
 //       strings. Each 2-Dimensional array element holds a separate
 //       string for error prefix and error context.
 //
-//       Information extracted from this object will be used to
-//       generate error prefix information in a new instance of
-//       ErrPrefixDto.
+//       Information extracted from this object will generate error
+//       prefix information used to overwrite and replace the error
+//       prefix and error context information in the input
+//       parameter, 'errPrefDto'.
 //
 //
 //  errorPrefStr        string