+++ errpref/zzzt_errpref_0002220_test.go
@@ -1562,6 +1562,72 @@ func TestErrPrefixDto_NewIEmpty_001600(t *testing.T) {
 
 }
 
+type testIBuilderErrPref struct {
+	errPrefInfo [][2]string
+}
+
+func (tIBuilder *testIBuilderErrPref) GetEPrefStrings() [][2]string {
+
+	if tIBuilder.errPrefInfo == nil {
+		return nil
+	}
+
+	lenTwoDArray := len(tIBuilder.errPrefInfo)
+
+	result := make([][2]string, lenTwoDArray)
+
+	if lenTwoDArray == 0 {
+		return result
+	}
+
+	copy(result, tIBuilder.errPrefInfo)
+
+	return result
+}
+
+func (tIBuilder *testIBuilderErrPref) SetEPrefStrings(twoDStrArray [][2]string) {
+
+	lenTwoDArray := len(twoDStrArray)
+
+	if lenTwoDArray == 0 {
+		return
+	}
+
+	tIBuilder.errPrefInfo = make([][2]string, lenTwoDArray)
+
+	copy(tIBuilder.errPrefInfo, twoDStrArray)
+
+}
+
+func (tIBuilder *testIBuilderErrPref) String() string {
+
+	str := ""
+
+	lenTwoDArray := len(tIBuilder.errPrefInfo)
+
+	if lenTwoDArray == 0 {
+		return str
+	}
+
+	for i := 0; i < lenTwoDArray; i++ {
+
+		if len(tIBuilder.errPrefInfo[i][0]) == 0 {
+			continue
+		}
+
+		str += tIBuilder.errPrefInfo[i][0]
+
+		if len(tIBuilder.errPrefInfo[i][1]) == 0 {
+			str += "\n"
+		} else {
+			str += " : " + tIBuilder.errPrefInfo[i][1] +
+				"\n"
+		}
+	}
+
+	return str
+}
+
 type testIStringerErrPref struct {
 	locString string
 }